refactor: step01 package 분리
